CS: Allow for local PHPCS config file
PHPCS supports looking for `phpcs.xml` and then `phpcs.xml.dist` files (in that order, along with leading-dot versions as well).

By having the main PHPCS configuration as a `.dist` (distributed) file, and adding the `phpcs.xml` file to the `.gitignore`, then a `phpcs.xml` file can be created locally, to override `phpcs.xml.dist`. This is good when wanting to test against a slightly different ruleset.