select and option added to allowed HTML
The get_allowed_wp_kses_html function was missing "select" and "option" which stripped the HTML when trying to add settings fields.

I noticed this when making an update I'm working on for the wpgraphql-cors repo which has its own tab of settings. Without this change, its not possible to create settings of the "select" type.